New NTSC shader by PlainOldPants (#627)
* Add Patchy NTSC

* Patchy NTSC - Improve Genesis/MegaDrive settings

Still matched 100% by eye

* Patchy NTSC - Last minute fixes

Moved patchy-mesen to the nes_raw_palette folder, and added some optional, unfinished gamut correction LUTs.
The LUT step is skipped by default and is clearly labelled as "work in progress" in the settings.
The purpose of including the gamut correction inside an NTSC shader is to encourage the use of more complex methods for gamut conversion. Up to this point, Grade has been the most popular option for that, yet Grade has some less favorable compromises due to the entire gamut conversion process being done in real time.

* Delete ntsc/patchy-mesen-raw-palette.slangp

* Patchy NTSC - Compress LUTs (except P22_80s_D65)

* Update P22-80s (not compressed)

All the LUTs that were compressed had also been updated prior to compressing. This one was updated too, but it was unable to be compressed. It was mistakenly omitted in the commit to update the LUTs.

* Remove notice about work-in-progress LUTs

Just ONE more thing